mastering-flask/requirements.txt
alembic==0.9.6
amqp==2.2.2
aniso8601==2.0.0
Babel==2.5.3
backports-abc==0.5
bcrypt==3.1.4
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
click==6.7
cssmin==0.2.0
dominate==2.3.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Celery-Helper==1.1.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
flask-mongoengine==0.9.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flower==0.9.2
ForgeryPy==0.1
futures==3.2.0
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.2
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
mongoengine==0.15.0
pycparser==2.18
pymongo==3.6.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
redis==2.10.6
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.15
tornado==4.5.3
vine==1.1.4
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.13
WTForms==2.1
